Finalized agent/server decoupling; started implementing attractor as a LocalAgent
