+++ server/agent/netagent.go
@@ -1,7 +1,11 @@
 package agent
 
 import (
+	"encoding/json"
 	"net"
+
+	"github.com/netgusto/bytearena/server/protocol"
+	"github.com/netgusto/bytearena/server/state"
 )
 
 type NetAgent interface {
@@ -23,6 +27,12 @@ func MakeNetAgentImp() NetAgentImp {
 
 func (agent NetAgentImp) String() string {
 	return "<NetAgentImp(" + agent.GetId().String() + ")>"
+}
+
+func (agent NetAgentImp) PutPerception(perception state.Perception, server protocol.AgentCommOperator) {
+	perceptionjson, _ := json.Marshal(perception)
+	message := []byte("{\"Method\": \"tick\", \"Arguments\": [0," + string(perceptionjson) + "]}\n") // TODO: remove 0 (ex turn)
+	server.GetNetworkCommServer().Send(message, agent.GetAddr())
 }
 
 func (agent NetAgentImp) SetAddr(addr net.Addr) NetAgent {
