+++ server/protocol/agentcommoperator.go
@@ -1,0 +1,7 @@
+package protocol
+
+// Interface to avoid circular dependencies between server and agent
+
+type AgentCommOperator interface {
+	GetNetworkCommServer() NetSender
+}
