+++ server/protocol/netsender.go
@@ -1,0 +1,7 @@
+package protocol
+
+import "net"
+
+type NetSender interface {
+	Send(message []byte, addr net.Addr)
+}
